Wandering Trader check
added check for wandering trader, logic like this:

wandering traders have 6 trades, and all of the apparently give 1 xp to it. So, just check if the entity has more than 4 trades and if all the trades give 1 xp, as villagers dont do that. I had to do this because I tried seeing what class the entity is and checking whether or not its a WanderingTraderEntity or something like that, but it always returns SimpleMerchant